Migrated importing code from index to wot_import.
